Return `Result<char, _>` from `EscapedStringIterator`s
While parsing the string we could come across issues which mean that it
is not well formed. Instead of quietly exiting as we have been doing, we
now return a `Result` from the iterator.

This slightly clunky element-by-element result can be turned into a
`Result` of string by using a `collect`.

Note taht we still only handle single-codepoint escaped charaters.