Rename JSONValue::parse -> JSONValue::load and add JSONValueType::Error
The function `parse` implied it did much more than what was actually
happening. In fact, it was only checking the first character of the
payload to guess the type of the string. The name `load` is closer to
what it actually does.

It is now possible to `load` a JSON which obviously is nonconformant
(i.e. its first character is wrong). These have a type
JSONValueType::Error.