feat: Add code to parse floats ourselves
The standard library float parser is very easy to use but is quite
extensive and careful about how it parses floats. This results in a
large amount of bloat. Further, it recognises strings that are not valid
JSON.

We can restrict to valid JSON and use a slightly more basic parser to
decrease the code footprint of the library.